Refactor: readline 분리 및 매직넘버 상수화 & view 로직 추가
input 사용자 입력 에러 핸들 & split 함수 분리

output 주석 제거 및 함수명 통일
